Support ICMP Network Policies via CRD
Network Policies support TCP, UDP and SCTP. However, there are use cases
where we need add ICMP to the mix too. The prposed changes allow defining
this via a new CRD of kind ExtNetworkPolicy. Currently this can only be
used to define ICMP v4 policies with ICMP types - this can be extended
as needed to include ICMP codes etc.

The reason to do this via CRD is to keep its use optional - with the
possibility of using this as the uber-Network Policy (in lieu of
using NetworkPolicy and ExtNetworkPolicy) in the future, if it makes
sense. The spec is kept as close as possible to NetworkPolicy since
ICMP policies have the same specification as TCP, UDP etc. (with ports
replaced with types). The idea is also to leverage existing implemtation
for NetworkPolicies, since both have some commonality, e.g. if
an ICMP policy is configured, there should be a default deny
others, just as we configure any NetworkPolicies today. Where
possible we make slight change to existing code to accommodate
both - but if that becomes restrictive, we can separate them out
in the future.

We define the CRD as v1alpha1 with the expectation, that this could
change with experience.

An example config (after deploying ExtNetworkPolicy CRD) looks like:

	apiVersion: k8s.ovn.org/v1alpha1
	kind: ExtNetworkPolicy
	metadata:
	  name: XXXext-policy
	  namespace: YYY
	spec:
	 podSelector:
	   matchLabels:
	     YYY-role: ZZZ
	 policyTypes:
	 - Ingress
	 ingress:
	  - from:
	    - podSelector:
	        matchLabels:
	          YYY-role: VVV
	    ports:
	    - protocol: ICMP
	  - from:
	    - podSelector:
	        matchLabels:
	          YYY-role: WWW
	    ports:
	    - protocol: ICMP
	      type: 8

Signed-off-by: Venu Iyer <venugopali@nvidia.com>